+++ src/main/java/ortus/boxlang/modules/esapi/util/KeyDirectory.java
@@ -40,6 +40,7 @@ public class KeyDirectory {
 	public static final Key	htmlAttribute			= Key.of( "htmlAttribute" );
 	public static final Key	javascript				= Key.of( "javascript" );
 	public static final Key	json					= Key.of( "json" );
+	public static final Key	directives				= Key.of( "directives" );
 	public static final Key	ldap					= Key.of( "ldap" );
 	public static final Key	moduleName				= Key.of( "moduleName" );
 	public static final Key	mysql					= Key.of( "mysql" );
@@ -48,6 +49,7 @@ public class KeyDirectory {
 	public static final Key	policy					= Key.of( "policy" );
 	public static final Key	restrictMultiple		= Key.of( "restrictMultiple" );
 	public static final Key	restrictMixed			= Key.of( "restrictMixed" );
+	public static final Key	throwOnError			= Key.of( "throwOnError" );
 	public static final Key	url						= Key.of( "url" );
 	public static final Key	urlEncode				= Key.of( "urlEncode" );
 	public static final Key	xml						= Key.of( "xml" );